# Permit skipping metadata for MariaDB