--- a/src/MySqlConnector/Core/CommandListPosition.cs
+++ b/src/MySqlConnector/Core/CommandListPosition.cs
@@ -8,6 +8,7 @@ internal struct CommandListPosition
 	public CommandListPosition(IReadOnlyList<IMySqlCommand> commands)
 	{
 		Commands = commands;
+		PreparedStatements = null;
 		CommandIndex = 0;
 		PreparedStatementIndex = 0;
 	}
@@ -18,6 +19,11 @@ internal struct CommandListPosition
 	public IReadOnlyList<IMySqlCommand> Commands { get; }
 
 	/// <summary>
+	/// Associated prepared statements of commands
+	/// </summary>
+	public PreparedStatements? PreparedStatements;
+
+	/// <summary>
 	/// The index of the current command.
 	/// </summary>
 	public int CommandIndex;
@@ -26,4 +32,9 @@ internal struct CommandListPosition
 	/// If the current command is a prepared statement, the index of the current prepared statement for that command.
 	/// </summary>
 	public int PreparedStatementIndex;
+
+	/// <summary>
+	/// Retrieve the last used prepared statement
+	/// </summary>
+	public PreparedStatement? LastUsedPreparedStatement;
 }