--- a/src/MySqlConnector/Core/PreparedStatement.cs
+++ b/src/MySqlConnector/Core/PreparedStatement.cs
@@ -17,6 +17,6 @@ internal sealed class PreparedStatement
 
 	public int StatementId { get; }
 	public ParsedStatement Statement { get; }
-	public ColumnDefinitionPayload[]? Columns { get; }
+	public ColumnDefinitionPayload[]? Columns { get; internal set; }
 	public ColumnDefinitionPayload[]? Parameters { get; }
 }