--- a/src/MySqlConnector/Protocol/ProtocolCapabilities.cs
+++ b/src/MySqlConnector/Protocol/ProtocolCapabilities.cs
@@ -147,4 +147,14 @@ internal enum ProtocolCapabilities : ulong
 	/// Support of array binding.
 	/// </summary>
 	MariaDbStatementBulkOperations = 0x4_0000_0000,
+
+	/// <summary>
+	/// metadata extended information
+	/// </summary>
+	MariaDbExtendedTypeInfo = 0x8_0000_0000,
+
+	/// <summary>
+	/// permit metadata caching
+	/// </summary>
+	MariaDbCacheMetadata = 0x16_0000_0000,
 }