Since MariaDB 10.6 (with https://jira.mariadb.org/browse/MDEV-19237) binary result-set can skip sending metadata when they haven't changed.
This avoid network data and parsing client side.

Benchmarks results using DDL:
```
CREATE TABLE test100 (i1 int,i2 int,...,i100 int);
INSERT INTO test100 value (1,2,...,100);
```

test methods:
```
    [Benchmark]
    public async Task<int> readOneRow100FieldsAsync()
    {
        var total = 0;
        using (var cmd = Connection.CreateCommand())
        {
            cmd.CommandText = "select * FROM test100";
            using (var reader = await cmd.ExecuteReaderAsync())
            {
                while (await reader.ReadAsync())
                    for (var i = 0; i < 100; i++)
                        total += reader.GetInt32(i);
            }
        }
        return total;
    }

    [Benchmark]
    public async Task<int> readOneRow100FieldsPrepareAsync()
    {
        var total = 0;
        using (var cmd = Connection.CreateCommand())
        {
            cmd.CommandText = "select * FROM test100";
            cmd.Prepare();
            using (var reader = await cmd.ExecuteReaderAsync())
            {
                while (await reader.ReadAsync())
                    for (var i = 0; i < 100; i++)
                        total += reader.GetInt32(i);
            }
        }
        return total;
    }
```

before PR:
```
|                          Method |        Library |      Mean |    Error |
|-------------------------------- |--------------- |----------:|---------:|
|        readOneRow100FieldsAsync | MySqlConnector | 104.47 us | 1.468 us |
| readOneRow100FieldsPrepareAsync | MySqlConnector |  92.49 us | 1.154 us |
```
After PR:
```
|                          Method |        Library |      Mean |    Error |
|-------------------------------- |--------------- |----------:|---------:|
|        readOneRow100FieldsAsync | MySqlConnector | 103.99 us | 0.914 us |
| readOneRow100FieldsPrepareAsync | MySqlConnector |  71.50 us | 1.402 us |
```