@bgrainger 2023-04-01
Thanks! I will take a look when I'm back from vacation.

I assume this may be related to https://github.com/mysql-net/MySqlConnector/issues/446 although MariaDB uses a different flag value to enable the capability? (Or maybe it's a completely different on-the-wire way to avoid sending metadata?)