@rusher 2023-04-03
In term of protocol, that correspond to the mysql change you indicate. 
Still this is not the exact same thing : 
* for MySQL there a capability CLIENT_OPTIONAL_RESULTSET_METADATA that permit to enable this functionality, then when @@session.resultset_metadata is "NONE", all result-set metadata won't then be sent.
* for MariaDB, when client set capability [MARIADB_CLIENT_CACHE_METADATA](https://mariadb.com/kb/en/connection/#capabilities), metadatas *MIGHT* not be send. We don't want the client to remember all metadata for each type of query. When MARIADB_CLIENT_CACHE_METADATA is set, server will not resent metadata to COM_STMT_EXECUTE result-set if metadata has not changed. So to resume, first execute result will have metadata, second will not.